refactor(dirt-files): relocated dirt related files to dirt folder
