+++ client/src/components/dirt/dirt-counter.tsx
