+++ client/src/components/dirt/dirt-overlay.tsx
