+++ client/src/components/dirt/dirt-spot.tsx
