+++ client/src/pages/game.tsx
@@ -14,8 +14,8 @@ import { motion } from 'framer-motion';
 import { useActiveAquarium } from '../store/active-aquarium';
 import { initialAquariums } from '@/data/mock-aquarium';
 import { useDirtSystemFixed as useDirtSystem } from '@/hooks/use-dirt-system-fixed';
-import { DirtOverlay } from '@/components/game/dirt-overlay';
-import { DirtDebugControls } from '@/components/game/dirt-debug-controls';
+import { DirtOverlay } from '@/components/dirt/dirt-overlay';
+import { DirtDebugControls } from '@/components/dirt/dirt-debug-controls';
 import { useFeedingSystem } from '@/systems/feeding-system';
 import { FeedingAquarium } from '@/components/game/feeding-aquarium';
 import { useAccount } from '@starknet-react/core';
